% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chart.R
\name{draw_chart}
\alias{draw_chart}
\title{Draw growth chart}
\usage{
draw_chart(chart, height = 29.7/2.54, width = 21/2.54)
}
\arguments{
\item{chart}{A \code{grob} or \code{gTree} object}

\item{height}{Height of pdf device (in inches)}

\item{width}{Width of pdf device (in inches)}
}
\value{
A grid object
}
\description{
Produce a pdf of the growth chart
}
\examples{
\dontrun{
z <- load_chart("NJAA")
draw_chart(z)
}
}
