% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rectangular.R
\name{extract_rectangular}
\alias{extract_rectangular}
\title{Extract rectangular plots from a chart}
\usage{
extract_rectangular(chartcode, ynames = c("hgt", "wgt", "hdc", "bmi",
  "wfh"), asp = NULL, width = unit(18, "cm"))
}
\arguments{
\item{chartcode}{The chart code}

\item{ynames}{Vector with names of measures to extract}

\item{asp}{Aspect ratio. Specify \code{asp = 1} for square
data area. The default takes the aspect ratio from the
chart.}

\item{width}{The width of the plot area + margins. The default
is \code{unit("18", "cm")}}
}
\description{
This function will extract the rectangular plots from the chart. It
will redefine the axes as rectangular, recalculate the gridlines,
extend the axis labels, and move around annotation object. This
version preserve the original aspect ratio.
}
\examples{
\dontrun{
library(grid)
library(chartbox)
p <- extract_rectangular("NJAA")
palette(palettes[1, ])
grid.newpage()
grid.draw(p[["hgt"]])
}
}
