% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chartcode.R
\name{get_chartcode}
\alias{get_chartcode}
\title{Construct the chart code from the user input}
\usage{
get_chartcode(chartgrp = c("nl2010", "preterm", "who"),
  etn = c("netherlands", "turkish", "moroccan", "hindustani"),
  sex = c("male", "female"), agegrp = c("0-15m", "0-4y", "1-21y",
  "0-21y", "0-4ya"), side = c("front", "back", "-hdc", "both"),
  week = 32, language = c("dutch", "english"), version = "3")
}
\arguments{
\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'} or \code{'who'}}

\item{etn}{Either \code{'netherlands'}, \code{'turkish'}, \code{'moroccan'} or
\code{'hindustani'}}

\item{sex}{Either \code{'male'} or \code{'female'}}

\item{agegrp}{Either \code{'0-15m'}, \code{'0-4y'}, \code{'1-21y'}, \code{'0-21y'}
or \code{'0-4ya'}. Age group \code{'0-4ya'} provides the 0-4 chart with weight
for age (design E).}

\item{side}{Either \code{'front'}, \code{'back'}, \code{'-hdc'} or \code{'both'}}

\item{week}{A number between 25 and 36 gestational age. Only used for preterm charts.}

\item{language}{Language: \code{'dutch'} or \code{'english'}}

\item{version}{Version number}
}
\description{
Construct the chart code from the user input
}
