% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viewport.R
\name{get_viewport}
\alias{get_viewport}
\title{Obtain the viewport number of the chart}
\usage{
get_viewport(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{chartbox::create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
An integer with the viewport number
}
\description{
Obtain the viewport number of the chart
}
