% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_charts.R
\name{list_charts}
\alias{list_charts}
\title{List available growth charts}
\usage{
list_charts(chartgrp = NULL)
}
\arguments{
\item{chartgrp}{The chart group}
}
\value{
A data frame
}
\description{
List available growth charts
}
\examples{
# list all available charts
list_charts()
}
