% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chart.R
\name{load_chart}
\alias{load_chart}
\title{Load growth chart}
\usage{
load_chart(chartcode)
}
\arguments{
\item{chartcode}{A string containing the chart code. If a vector
only the first element is processed.}
}
\value{
A grid object. Return \code{NULL} if the chart cannot
be loaded.
}
\description{
Load growth chart
}
\examples{
z <- load_chart("NJAA")
}
