% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Palettes for Growth Charts}
\format{A data frame with 10 observations 6 variables
\describe{ 
\item{bg}{Background color}
\item{fg}{Foreground color} 
\item{dark}{Blackish color} 
\item{light}{Lightish color}
\item{aux1}{Color aux1}
\item{aux2}{Color aux2} }}
\description{
Palettes for Growth Charts
}
\seealso{
\code{\link[chaco]{create.palettes}}
}
\keyword{datasets}
