% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chartgrp.R
\name{parse_chartgrp}
\alias{parse_chartgrp}
\title{Parses the chart group}
\usage{
parse_chartgrp(chartcode = NULL)
}
\arguments{
\item{chartcode}{the chart code, usually constructed by \code{get_chartcode()}}
}
\value{
The function returns a string with the chart group
}
\description{
Parses the chart group
}
