% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_chart.R
\name{pdf_chart}
\alias{pdf_chart}
\title{Save growth chart as pdf}
\usage{
pdf_chart(chart, height = 29.7/2.54, width = 21/2.54)
}
\arguments{
\item{chart}{A \code{grob} or \code{gTree} object}

\item{height}{Height of pdf device (in inches)}

\item{width}{Width of pdf device (in inches)}
}
\value{
A grid object
}
\description{
Produce a pdf of the growth chart
}
\examples{
\dontrun{
z <- load_chart("NJAA")
pdf_chart(z)
}
}
