% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ynames_lookup.R
\docType{data}
\name{ynames_lookup}
\alias{ynames_lookup}
\title{Details of installed charts}
\format{An object of class \code{data.frame} with 212 rows and 8 columns.}
\usage{
ynames_lookup
}
\description{
A data set containing the kind of outcomes present on each chart
in \code{inst/library}
}
\keyword{datasets}
