% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viewport.R
\name{get_viewport}
\alias{get_viewport}
\title{Obtain the viewport number of the chart}
\usage{
get_viewport(chartgrp, chartcode, yname)
}
\arguments{
\item{chartgrp}{Chart group}

\item{chartcode}{The chart code, returned by \code{chartbox::create_chartcode()}}

\item{yname}{Names of the response variable}
}
\value{
An integer with the viewport number
}
\description{
Obtain the viewport number of the chart
}
