% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Palettes for Growth Charts}
\format{
A data frame with 10 observations 6 variables
\describe{
\item{bg}{Background color: Always \code{#F7F7F7} (\code{grey97})}
\item{fg}{Foreground color. The major chart color that
indicates the population. Used for lines and smaller fills.}
\item{dark}{Blackish color for text and labels.}
\item{light}{Light color used for filling the normal centile range.}
\item{aux1}{Color aux1.}
\item{aux2}{Color aux2}}
}
\description{
Palettes for Growth Charts
}
\keyword{datasets}
